requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.0.9

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Django REST framework
djangorestframework==3.11.0
django-filter

# JWT
pyjwt==2.1.0

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1
tornado>=5.0.0,<6.0.0
flower==0.9.2
